feat(functionality) Code Test Answer
I only have 1 commit for this trial, as I'm not really
used to pushing commits which I am not really sure
that are correct and working.

On the otherhand, provided are the solutions I was
able to come up with, in referenced to the material,
explained via the README file. To demonstrate each
of my answers, I have prepared to the script to
accept command line arguments.

- To demonstrate my solution for #1 run:
node index async

- To demonstrate my solution for #2 run:
node index stream

- To demonstrate my solution for #3 run:
node index resource